\name{plot.linkage.all}
\alias{plot.linkage.all}
\title{
  Plot a linkage.all object
}
\description{
Plot a linkage.all object
}
\usage{
plot.linkage.all(x, ...)
}
\arguments{
\item{x}{
A linkage.all object
}
\item{\dots }{
additional plot parameters, this method checks for 'title'
}
}
\value{
nothing is returned
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{
\code{\link{ibdreg}},
\code{\link{linkage.all}},
\code{\link{print.linkage.all}}
}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
