% $Author: sinnwell $ -->
% $Date: 2010/02/08 18:19:59 $ -->
% $Header: /people/biostat3/sinnwell/Projects/IBDReg/Make/RCS/plot.linkage.all.Rd,v 1.3 2010/02/08 18:19:59 sinnwell Exp $ -->
% $Locker:  $ -->
%
%$Log: plot.linkage.all.Rd,v $
%Revision 1.3  2010/02/08 18:19:59  sinnwell
%rm empty sections
%
%Revision 1.2  2008/10/15 16:28:17  sinnwell
%usage
%
%Revision 1.1  2008/10/14 20:43:33  sinnwell
%Initial revision
%
% Revision 1.4  2008/10/14 20:30:16  sinnwell
% addkeywords
%

\name{plot.linkage.all}
\alias{plot.linkage.all}
\title{
  Plot a linkage.all object
}
\description{
Plot a linkage.all object
}
\usage{
\method{plot}{linkage.all}(x, ...)
}
\arguments{
\item{x}{
A linkage.all object
}
\item{\dots }{
additional plot parameters, this method checks for 'title'
}
}
\value{
nothing is returned
}
\seealso{
\code{\link{ibdreg}},
\code{\link{linkage.all}},
\code{\link{print.linkage.all}}
}
% docclass is function
% Converted by Sd2Rd version 43267.
