\name{print.linkage.all}
\alias{print.linkage.all}
\title{
  Print or summarize a linkage.all object
}
\description{
Print a linkage.all object, made within ibdreg. For the linkage test,
print the test statistic, degrees of freedom, and pvalue at the
position with the smallest pvalue.
}
\usage{
print.linkage.all(x, digits=max(options()$digits - 2, 5), ...)
}
\arguments{

\item{x}{
A linkage.all object
}
\item{digits}{
number of significant digits to print for numeric values
}
\item{\dots }{
additional display parameters
}
}
\value{
nothing is returned
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{
\code{\link{ibdreg}},
\code{\link{linkage.all}},
\code{\link{plot.linkage.all}}
}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
